+++ src/GRA.Domain.Model/Filters/AvatarFilter.cs
@@ -6,7 +6,10 @@ namespace GRA.Domain.Model.Filters
 {
     public class AvatarFilter : BaseFilter
     {
-        public bool? Unlockable { get; set; }
+        public bool Available { get; set; }
+        public bool Unavailable { get; set; }
+        public bool Unlockable { get; set; }
+        public bool CanBeUnlocked { get; set; }
         public ICollection<int> ItemIds { get; set; }
         public int? LayerId { get; set; }